// src/payload-types.ts
export interface SiteSetting {
  id: number;
  siteName: string;
  doctorName: string;
  clinicType: string;
  tagline?: string | null;
  domain: string;
  updatedAt?: string | null;
  createdAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "contact-information".
 */
export interface ContactInformation {
  id: number;
  phone: string;
  email?: string | null;
  address: string;
  mapURL?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "social-media".
 */
export interface SocialMedia {
  id: number;
  facebookURL?: string | null;
  instagramURL?: string | null;
  youtubeURL?: string | null;
  tiktokURL?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "clinic-timings".
 */
export interface ClinicTiming {
  id: number;
  schedule?:
    | {
        days: string;
        hours: string;
        closed?: boolean | null;
        id?: string | null;
      }[]
    | null;
  appointmentNote?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}